refactor: move thread runtime launch config
